H.W

+++ 03 - Programming Using C++ - Level 1/Age Variable H.W/Age Variable H.W.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main()
 {
-    int Age = 25;
+    short Age = 25;
 
     cout << "After 5 years you will be " << Age + 5 << " years old.\n";
